Change lifter arg type, provid natural modules
This commit changes the accepted type of
the lifter's lift_insns function to be exactly
identical to Disassembler's insns's return type.
The result is slightly smoother code with less
fanagling with the type system, which may seem
puny, but is good when you write a lot of
different applications that all use the lifter
interface.

In addition, some minor changes include adding
ocamlbuild/oasis and opam integration so that
testing could be done. Other noise includes
.merlin, gitignore and Makefile changes.